git: Extract GitStepMixin
